feat: New name mixin
